requirements.txt
lmdb
Pillow
tqdm
six